Show the category in a separate post. Update info DataBase

+++ admin/modules/blog/edit.php
@@ -1,6 +1,6 @@
 <?php
 
-$cats = R::find('categories', 'ORDER BY title ASC');
+$cats = R::find('categories', 'ORDER BY cat_title ASC');
 
 if (isset($_POST['postEdit'])) {
 
@@ -22,16 +22,18 @@ if (isset($_POST['postEdit'])) {
         $post->content = $_POST['content'];
         $post->editTime = time();
 
-
         // Удаление обложки
         if (isset($_POST['delete-cover']) && $_POST['delete-cover'] == 'on') {
 
             // Удалить файлы обложки
             $coverFolderLocation = ROOT . 'usercontent/blog/';
-            // echo $coverFolderLocation . $post->cover_small;
-            // die();
-            unlink($coverFolderLocation . $post->cover);
-            unlink($coverFolderLocation . $post->cover_small);
+
+            if (file_exists(ROOT . 'usercontent/blog/' . $post->cover) && !empty($user->cover)) {
+                unlink(ROOT . 'usercontent/blog/' . $post->cover);
+            }
+            if (file_exists(ROOT . 'usercontent/blog/' . $post->coverSmall) && !empty($user->coverSmall)) {
+                unlink(ROOT . 'usercontent/blog/' . $post->coverSmall);
+            }
 
             // Удалить записи в БД
             $post->cover = NULL;
@@ -39,18 +41,25 @@ if (isset($_POST['postEdit'])) {
         }
 
         // Если передано изображение - уменьшаем, сохраняем, записываем в БД
-        $coverFileName = saveUploadedImg('cover', [600, 300], 12, 'blog', [1110, 460], [290, 230]);
+        if (isset($_FILES['cover']['name']) && $_FILES['cover']['tmp_name'] !== '') {
+            // Обрабатываем картинку, сохраняем, и получаем имя файла
+            $coverFileName = saveUploadedImg('cover', [600, 300], 12, 'blog', [1110, 460], [290, 230]);
 
-        // Если новое изображение успешно загружено тогда удаляем старое
-        if ($coverFileName) {
-            // Удаляем старое изображение
-            unlink(ROOT . 'usercontent/blog/' . $post->cover);
-            unlink(ROOT . 'usercontent/blog/' . $post->coverSmall);
+            // Если новое изображение успешно загружено тогда удаляем старое
+            if ($coverFileName) {
+                // Удаляем старое изображение
+                if ( file_exists(ROOT . 'usercontent/blog/' . $post->cover) && !empty($user->cover)){
+                    unlink(ROOT . 'usercontent/blog/' . $post->cover);
+                }
+                if ( file_exists(ROOT . 'usercontent/blog/' . $post->coverSmall) && !empty($user->coverSmall)) {
+                    unlink(ROOT . 'usercontent/blog/' . $post->coverSmall);
+                }
+            }
+
+            // Сохраняем имя файла в БД
+            $post->cover = $coverFileName[0];
+            $post->coverSmall = $coverFileName[1];
         }
-
-        // Сохраняем имя файла в БД
-        $post->cover = $coverFileName[0];
-        $post->coverSmall = $coverFileName[1];
 
         R::store($post);
         $_SESSION['success'][] = ['title' => 'Пост успешно обновлен'];
